Refactor tests and controller
