+++ spec/support/features/checkouts_helper.rb
@@ -1,0 +1,13 @@
+module Features
+  def have_checked_out_book(title)
+    have_css('ul.books.unavailable li', text: title)
+  end
+
+  def have_returned_book(title)
+    have_css('ul.books.available li', text: title)
+  end
+
+  def have_available_book(title)
+    have_css('ul.books.available li', text: title)
+  end
+end
